[fix/#142] 빌드 오류 최종 수정 및 기타 오류 수정
