+++ app/(base)/arenas/components/DebatingArenaSection.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import useArenas from "@/hooks/useArenas";
 import ArenaSectionHeader from "./ArenaSectionHeader";
 import DebatingArenaCard from "./DebatingArenaCard";
