+++ app/(base)/arenas/components/WaitingArenaSection.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import useArenas from "@/hooks/useArenas";
 import ArenaSectionHeader from "./ArenaSectionHeader";
 import WaitingArenaCard from "./WaitingArenaCard";
