+++ app/(base)/profile/components/PointHistoryCard.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import Image from "next/image";
 import clsx from "clsx";
