+++ app/components/Footer.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 export default function Footer() {
     return (
         <footer className="w-full bg-background-100 py-8 flex flex-col items-center justify-center text-center font-sans">
